Refactor to support unit driven deploy process
Each of the build steps are now in their own separate target making it
easier to test each step.  In the process, I also added a destroy-stack
target to support the clean purge of the stack from the target account.

I realize calling this commit a *refactor* is very loose, as there is
a significant amount that was changed. However, in defense, the
functionality of the Makefile remains the same.